fix: animation and function refactor
